requirements.txt
gradio
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
pymupdf
python-docx
requests
python-dotenv
transformers
torch

